+++ src/response/session/session.rs
@@ -3,28 +3,28 @@ use crate::SessionStack;
 use std::time::Duration;
 
 // NOTE: u16 mb not enough!
-pub const SESSION_ACTIVITY: usize = 1;
-pub const SESSION_ALERTS: usize = 1 << 1;
-pub const SESSION_ATTACHED: usize = 1 << 2;
-pub const SESSION_CREATED: usize = 1 << 3;
-pub const SESSION_FORMAT: usize = 1 << 4;
-pub const SESSION_GROUP: usize = 1 << 5;
-pub const SESSION_GROUP_LIST: usize = 1 << 6;
-pub const SESSION_GROUP_SIZE: usize = 1 << 7;
-pub const SESSION_GROUPED: usize = 1 << 8;
-pub const SESSION_ID: usize = 1 << 9;
-pub const SESSION_LAST_ATTACHED: usize = 1 << 10;
-pub const SESSION_MANY_ATTACHED: usize = 1 << 11;
-pub const SESSION_NAME: usize = 1 << 12;
-pub const SESSION_STACK: usize = 1 << 13;
-pub const SESSION_WINDOWS: usize = 1 << 14;
+pub const SESSION_ACTIVITY: u16 = 1;
+pub const SESSION_ALERTS: u16 = 1 << 1;
+pub const SESSION_ATTACHED: u16 = 1 << 2;
+pub const SESSION_CREATED: u16 = 1 << 3;
+pub const SESSION_FORMAT: u16 = 1 << 4;
+pub const SESSION_GROUP: u16 = 1 << 5;
+pub const SESSION_GROUP_LIST: u16 = 1 << 6;
+pub const SESSION_GROUP_SIZE: u16 = 1 << 7;
+pub const SESSION_GROUPED: u16 = 1 << 8;
+pub const SESSION_ID: u16 = 1 << 9;
+pub const SESSION_LAST_ATTACHED: u16 = 1 << 10;
+pub const SESSION_MANY_ATTACHED: u16 = 1 << 11;
+pub const SESSION_NAME: u16 = 1 << 12;
+pub const SESSION_STACK: u16 = 1 << 13;
+pub const SESSION_WINDOWS: u16 = 1 << 14;
 
 // number of all flags, needed for array init
 pub const SESSION_FLAGS_NUM: usize = 15;
 
-pub const SESSION_NONE: usize = 0;
+pub const SESSION_NONE: u16 = 0;
 //pub const SESSION_DEFAULT: usize = SESSION_ID | SESSION_NAME;
-pub const SESSION_ALL: usize = SESSION_ACTIVITY
+pub const SESSION_ALL: u16 = SESSION_ACTIVITY
     | SESSION_ATTACHED
     | SESSION_ALERTS
     | SESSION_CREATED
@@ -44,7 +44,7 @@ pub const SESSION_VARS_SEPARATOR: &str = ":";
 
 // FIXME: regex name can be anything, and other keys should be checked better
 // NOTE: no colons or periods (ref: int session_check_name(const char *name))
-pub const SESSION_VARS: [(&str, usize, fn(s: &mut Session, p: &str)); SESSION_FLAGS_NUM] = [
+pub const SESSION_VARS: [(&str, u16, fn(s: &mut Session, p: &str)); SESSION_FLAGS_NUM] = [
     ("session_activity", SESSION_ACTIVITY, |s, p| {
         s.activity = p.parse().ok().map(Duration::from_millis)
     }),
@@ -129,7 +129,7 @@ impl Session {
         Default::default()
     }
 
-    pub fn from_str(s: &str, bitflags: usize) -> Result<Self, Error> {
+    pub fn from_str(s: &str, bitflags: u16) -> Result<Self, Error> {
         let sv: Vec<&str> = s.split(SESSION_VARS_SEPARATOR).collect();
         let mut sv = sv.iter();
         // XXX: optimize?
